% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/foodwebfunctions.R
\name{mean.r}
\alias{mean.r}
\title{calculates the mean of a vector while excluding all NAs from it (roxygen comment)}
\usage{
\method{mean}{r}(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
calculates the mean of a vector while excluding all NAs from it (roxygen comment)
}

