% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web_generator_r.R
\name{web_generator_r}
\alias{web_generator_r}
\title{Create a predator/prey table}
\usage{
web_generator_r(spe_list, f_web)
}
\arguments{
\item{spe_list}{is a string vector containing a list of species}

\item{f_web}{is a data frame containing predator/prey relationships}
}
\description{
Like web_generator it creates a table of predator and corresponding prey.
It it faster but it gives a different result and may not make any sense! To discuss with the group
}

