% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web_generator.R
\name{web_generator}
\alias{web_generator}
\title{creates a food web from predator/prey relationships}
\usage{
web_generator(spe_list, f_web)
}
\arguments{
\item{spe_list}{is a string vector containing a list of species}

\item{f_web}{is a data frame containing predator/prey relationships}
}
\description{
creates a food web from predator/prey relationships
}

