% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web_generator.R
\name{web_generator_r}
\alias{web_generator_r}
\title{Creates a predator/prey table}
\usage{
web_generator_r(spe_list, f_web)
}
\arguments{
\item{spe_list}{is a string vector containing a list of species}

\item{f_web}{is a data frame containing predator/prey relationships}
}
\description{
Creates a table in which each row contains a predator and one
of its prey. Each predator has as many entries as it has different prey.
}

